Fix clock adjustment when there's more than ~140 leds.

+++ clockless.h
@@ -110,7 +110,7 @@ public:
 		showRGBInternal<0, false>(nLeds, scale, (const byte*)&data);
 
 		// Adjust the timer
-		long microsTaken = CLKS_TO_MICROS(nLeds * 24 * (T1 + T2 + T3));
+		long microsTaken = CLKS_TO_MICROS((long)nLeds * 24 * (T1 + T2 + T3));
 		MS_COUNTER += (microsTaken / 1000);
 		sei();
 		mWait.mark();
@@ -123,7 +123,7 @@ public:
 		showRGBInternal<0, true>(nLeds, scale, (const byte*)rgbdata);
 
 		// Adjust the timer
-		long microsTaken = CLKS_TO_MICROS(nLeds * 24 * (T1 + T2 + T3));
+		long microsTaken = CLKS_TO_MICROS((long)nLeds * 24 * (T1 + T2 + T3));
 		MS_COUNTER += (microsTaken / 1000);
 		sei();
 		mWait.mark();
@@ -137,7 +137,7 @@ public:
 		showRGBInternal<1, true>(nLeds, scale, (const byte*)rgbdata);
 
 		// Adjust the timer
-		long microsTaken = CLKS_TO_MICROS(nLeds * 24 * (T1 + T2 + T3));
+		long microsTaken = CLKS_TO_MICROS((long)nLeds * 24 * (T1 + T2 + T3));
 		MS_COUNTER += (microsTaken / 1000);
 		sei();
 		mWait.mark();
